requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.3
    # via twitchapi
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via aiohttp
beautifulsoup4==4.11.1
    # via bs4
bottle==0.12.23
    # via
    #   bottle-websocket
    #   eel
bottle-websocket==0.2.9
    # via eel
brotli==1.0.9
    # via yt-dlp
bs4==0.0.1
    # via -r requirements.in
certifi==2022.9.24
    # via
    #   requests
    #   yt-dlp
cffi==1.15.1
    # via
    #   clr-loader
    #   gevent
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via gtts
clr-loader==0.2.4
    # via pythonnet
colorama==0.4.6
    # via click
coverpy==0.0.4
    # via -r requirements.in
decorator==5.1.1
    # via validators
discord-webhook==1.0.0
    # via -r requirements.in
eel==0.15.0
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
future==0.18.2
    # via eel
gevent==22.10.2
    # via gevent-websocket
gevent-websocket==0.10.1
    # via bottle-websocket
greenlet==2.0.1
    # via gevent
gtts==2.3.0
    # via -r requirements.in
idna==3.4
    # via
    #   requests
    #   yarl
keyboard==0.13.5
    # via -r requirements.in
multidict==6.0.3
    # via
    #   aiohttp
    #   yarl
mutagen==1.46.0
    # via yt-dlp
obsws-python==1.3.0
    # via -r requirements.in
pillow==9.3.0
    # via -r requirements.in
proxy-tools==0.1.0
    # via pywebview
pycparser==2.21
    # via cffi
pycryptodomex==3.16.0
    # via yt-dlp
pyparsing==3.0.9
    # via eel
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   twitchapi
python-dotenv==0.21.0
    # via -r requirements.in
pythonnet==3.0.1
    # via pywebview
pytube==12.1.0
    # via -r requirements.in
pywebview==3.7.2
    # via -r requirements.in
requests==2.28.1
    # via
    #   -r requirements.in
    #   coverpy
    #   discord-webhook
    #   gtts
    #   twitchapi
rx==3.2.0
    # via -r requirements.in
six==1.16.0
    # via
    #   gtts
    #   python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
tomli==2.0.1
    # via obsws-python
twitchapi==2.5.7
    # via -r requirements.in
typing-extensions==4.4.0
    # via twitchapi
urllib3==1.26.13
    # via requests
validators==0.20.0
    # via -r requirements.in
websocket-client==1.4.2
    # via obsws-python
websockets==10.4
    # via
    #   twitchapi
    #   yt-dlp
whichcraft==0.6.1
    # via eel
yarl==1.8.2
    # via aiohttp
yt-dlp==2022.11.11
    # via -r requirements.in
zope-event==4.5.0
    # via gevent
zope-interface==5.5.2
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools

akeyless~=2.20.3
pygame~=2.1.3.dev8
screeninfo~=0.8.1